edited pytest APIs
